% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixTCRclean-package.R
\docType{package}
\name{MixTCRclean.package}
\alias{MixTCRclean-package}
\alias{MixTCRclean.package}
\title{MixTCRclean: Clean TCR data}
\description{
Clean TCR data by checking V/J names, CDR3 length and concordance with germline sequences.
}
\author{
\strong{Maintainer}: David Gfeller \email{david.gfeller@unil.ch} (\href{https://orcid.org/0000-0002-3952-0930}{ORCID})

Authors:
\itemize{
  \item Dana Moreno \email{dana.moreno@unil.ch} (\href{https://orcid.org/0009-0007-9758-3335}{ORCID})
}

}
\keyword{internal}
