% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixTCRclean.R
\name{MixTCRclean}
\alias{MixTCRclean}
\title{MixTCRclean: Predict inconsistencies between V/J genes and CDR3}
\usage{
MixTCRclean(
  input,
  output.path = NULL,
  chain = "AB",
  use.allele = F,
  correct.gene.names = T,
  use.mouse.strain = F,
  check.cdr3.mode = 1,
  start.lg = 1,
  end.lg = 2,
  filename.output = NULL,
  species.default = "HomoSapiens",
  verbose = 1,
  build.clones = F,
  keep.incomplete.chain = T,
  seq.protocol = "Default",
  keep.colnames.origin = F
)
}
\arguments{
\item{input}{Either a .csv, .txt, .tsv file or a data.frame or a list in the MixTCRclean format.
Columns should ideally include 'TRAV','TRAJ','cdr3_TRA','TRBV','TRBJ','cdr3_TRB'.}

\item{output.path}{Name of the output directory (default=NULL). Created if it doesn't exist.}

\item{chain}{Which chain(s) to check for inconsistencies. Options: "AB" (default), "A", or "B".}

\item{use.allele}{Whether to keep allele-level resolution. Default is FALSE (merge alleles).}

\item{correct.gene.names}{Whether to auto-correct V/J names using an internal dictionary. Default is TRUE.}

\item{use.mouse.strain}{Whether to treat mouse strain-specific genes separately. Default is FALSE.}

\item{check.cdr3.mode}{Mode for filtering incompatible CDR3 sequences. 0 = none, 1 = default QC, 2 = use germline conservation thresholds derived from statistical modelling}

\item{start.lg, end.lg}{CDR3 positions or lengths to use for motif building or analysis. Defaults are 1 and 2.}

\item{filename.output}{Custom filename for outputs. Default = NULL.}

\item{species.default}{Default species used if not in input. Either "HomoSapiens" or "MusMusculus". Default is "HomoSapiens".}

\item{verbose}{Verbosity level in the console. 0 = silent, 1 = basic info (default), 2 = detailed, 3 = full.}

\item{build.clones}{Whether to reconstruct clones from clone.id or complex.id fields. Default is FALSE.}

\item{keep.incomplete.chain}{Whether to retain incomplete alpha/beta chains. Default is TRUE.}

\item{seq.protocol}{Sequencing protocol. Options: "Default" or "SEQTR". Affects gene merging.}

\item{keep.colnames.origin}{Whether to preserve original column names. Default is FALSE.}
}
\value{
the input dataframe with additional columns for the predicted consistency status and comments (TRA_pred_consistent, Comments_TRA, TRB_pred_consistent, Comments_TRB)
}
\description{
MixTCRclean is an R package that detects inconsistencies between V/J gene assignments and CDR3 amino acid sequences in TCR repertoire datasets.
MixTCRclean checks the beginning and end of CDR3 sequences for the expected germline amino acids based on the germline conservation thresholds derived from statistical modelling of a large TCR repertoire datasets
}
