% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixTCRclean.R
\name{MixTCRclean}
\alias{MixTCRclean}
\title{MixTCRclean: Plot TCR motifs}
\usage{
MixTCRclean(
  input1,
  output.path = NULL,
  chain = "AB",
  use.allele = F,
  correct.gene.names = T,
  use.mouse.strain = F,
  check.cdr3.mode = 1,
  start.lg = 1,
  end.lg = 2,
  filename.output = NULL,
  set.cdr3a.length = NA,
  set.cdr3b.length = NA,
  species.default = "HomoSapiens",
  verbose = 1,
  build.clones = F,
  keep.incomplete.chain = T,
  seq.protocol = "Default",
  keep.colnames.origin = F
)
}
\arguments{
\item{input1}{Either a .csv, .txt, .tsv file or a data.frame or a list in the MixTCRclean format.
Columns should ideally include 'TRAV','TRAJ','cdr3_TRA','TRBV','TRBJ','cdr3_TRB'.
If TCRs from multiple sources are included, the 'model' and 'species' columns are recommended.}

\item{output.path}{Name of the output directory (default=NULL). Created if it doesn't exist.}

\item{chain}{Which chain(s) to plot. Options: "AB" (default), "A", or "B".}

\item{use.allele}{Whether to keep allele-level resolution. Default is FALSE (merge alleles).}

\item{correct.gene.names}{Whether to auto-correct V/J names using an internal dictionary. Default is TRUE.}

\item{use.mouse.strain}{Whether to treat mouse strain-specific genes separately. Default is FALSE.}

\item{check.cdr3.mode}{Mode for filtering incompatible CDR3 sequences. 0 = none, 1 = default QC.}

\item{start.lg, }{end.lg CDR3 positions or lengths to use for motif building or analysis. Defaults are 1 and 2.}

\item{filename.output}{Custom filename for outputs. Default = NULL, filenames based on \code{model}.}

\item{set.cdr3a.length, set.cdr3b.length}{Length to use for motif visualization. Default = NA (auto-detect).}

\item{species.default}{Default species used if not in input1. Either "HomoSapiens" or "MusMusculus". Default is "HomoSapiens".}

\item{verbose}{Verbosity level in the console. 0 = silent, 1 = basic info (default), 2 = detailed, 3 = full.}

\item{build.clones}{Whether to reconstruct clones from clone.id or complex.id fields. Default is FALSE.}

\item{keep.incomplete.chain}{Whether to retain incomplete alpha/beta chains. Default is TRUE.}

\item{seq.protocol}{Sequencing protocol. Options: "Default" or "SEQTR". Affects gene merging.}

\item{keep.colnames.origin}{Whether to preserve original column names. Default is FALSE.}
}
\value{
A list containing plots (if enabled), processed data, and model statistics.
}
\description{
MixTCRclean is an R package to display TCR binding motifs for a set of TCRs provided by the user.
Typically, the input TCRs correspond to TCRs binding to a specific epitope, or isolated in a specific experiment.
MixTCRclean compares V usage, J usage, CDR3 length distribution and CDR3 sequence motifs to those expected from baseline TCR repertoire.
}
